# Optimize segmentation/rank.py for performance and maintainability