## Summary
This PR partially addresses #862 

> 🎯 **Purpose**: Improve performance of public functions in segmentation/rank.py file
>

**[ ✏️ Write your summary here. ]**
For memory I used the memory-profiler library. The code I used for benchmarking is copied below. In addition I sorted the imports and fixed the return type of the _issues_from_scores_ function as it always returns a numpy array.

**Code Setup**
```python
import random

import numpy as np

from cleanlab.segmentation.rank import get_label_quality_scores, issues_from_scores

SIZE = 1_000
DATASET_EXP_SIZE = 5
np.random.seed(0)
%load_ext memory_profiler

# Copied from the test file with minor changes
def generate_three_image_dataset(bad_index):
    good_gt = np.zeros((SIZE, SIZE))
    good_gt[:SIZE // 2, :] = 1.0
    bad_gt = np.ones((SIZE, SIZE))
    bad_gt[:SIZE // 2, :] = 0.0
    good_pr = np.random.random((2, SIZE, SIZE))
    good_pr[0, :SIZE // 2, :] = good_pr[0, :SIZE // 2, :] / 10
    good_pr[1, SIZE // 2:, :] = good_pr[1, SIZE // 2:, :] / 10

    val = np.binary_repr([4, 2, 1][bad_index], width=3)
    error = [int(case) for case in val]

    labels = []
    pred = []
    for case in val:
        if case == "0":
            labels.append(good_gt)
            pred.append(good_pr)
        else:
            labels.append(bad_gt)
            pred.append(good_pr)

    labels = np.array(labels)
    pred_probs = np.array(pred)
    return labels, pred_probs, error

# Create input data
labels, pred_probs, error = generate_three_image_dataset(random.randint(0, 2))
for _ in range(DATASET_EXP_SIZE):
    labels = np.append(labels, labels, axis=0)
    pred_probs = np.append(pred_probs, pred_probs, axis=0)

labels, pred_probs = labels.astype(int), pred_probs.astype(float)
image_scores_softmin, pixel_scores = get_label_quality_scores(labels, pred_probs, method="softmin", verbose=False)
```


**Current version**

```python
%timeit get_label_quality_scores(labels, pred_probs, method="softmin", verbose=False)
# 1.53 s ± 64.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
%memit get_label_quality_scores(labels, pred_probs, method="softmin", verbose=False);
# peak memory: 4592.04 MiB, increment: 1413.73 MiB
%memit issues_from_scores(image_scores_softmin, pixel_scores, threshold=0.5);
# peak memory: 3298.25 MiB, increment: 171.11 MiB
%timeit issues_from_scores(image_scores_softmin, pixel_scores, threshold=0.5);
# 346 ms ± 9.55 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
**This PR**
```python
%timeit get_label_quality_scores(labels, pred_probs, method="softmin", verbose=False)
# 1.43 s ± 61.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
%memit get_label_quality_scores(labels, pred_probs, method="softmin", verbose=False);
# peak memory: 3891.34 MiB, increment: 764.56 MiB
%memit issues_from_scores(image_scores_softmin, pixel_scores, threshold=0.5);
# peak memory: 3127.20 MiB, increment: 0.00 MiB
%timeit issues_from_scores(image_scores_softmin, pixel_scores, threshold=0.5);
# 51.5 ms ± 958 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

## Testing

> 🔍 Testing Done: Existing tests.


## References
A significant part of the time spent running the function _get_label_quality_scores_ is calling the softmax function. In the future I would like to keep working on #862 although I see little room for improvement in the softmax function.

## Reviewer Notes
This is my first PR to this repo.
> 💡 Include any specific points for the reviewer to consider during their review.